+++ Source_Code/Project_2-master/Project_2-master/database/migrations/2022_03_15_164730_add_programs_to_users_table.php
@@ -14,8 +14,10 @@ class AddProgramsToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('program_id')->nullable();
-            $table->foreign('program_id')->references('id')->on('programs')->onDelete('cascade'); // use this for field after specific column.
+            if (!Schema::hasColumn('users', 'program_id')) {
+                $table->unsignedBigInteger('program_id')->nullable();
+                $table->foreign('program_id')->references('id')->on('programs')->onDelete('cascade');
+            }
         });
     }
